requirements.txt
fastapi
uvicorn
pydantic
numpy
shapely
python-dotenv
earthengine-api
scikit-learn
